Update vehicles page to list and support filters
