+++ backups/backup-2025-09-22T20-48-01-334Z/manufacturer.json
@@ -5,7 +5,7 @@
     "_rev": "ZqU99J9YlRX1csb8eKiXYq",
     "_type": "manufacturer",
     "_updatedAt": "2025-09-22T19:06:10Z",
-    "description": "Ford is a leader in passenger and commercial vehicle manufacturers.",
+    "description": "Ford is a leader in passenger and utility vehicle manufacturers.",
     "name": "Ford",
     "slug": {
       "_type": "slug",
